src/main/java/org/example/hacaton/service/impl/UserServiceImpl.java
package org.example.hacaton.service.impl;


import lombok.RequiredArgsConstructor;
import org.example.hacaton.model.user.User;
import org.example.hacaton.repository.UserRepository;
import org.example.hacaton.service.UserService;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;
import org.example.hacaton.exception.CustomException;
import org.springframework.transaction.annotation.Transactional;

@Service
@RequiredArgsConstructor
public class UserServiceImpl implements UserService {

    private final UserRepository userRepository;

    @Override
    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
        return userRepository.findByEmail(email)
                .orElseThrow(() -> new UsernameNotFoundException("Пользователь с таким email не найден"));
    }

    @Override
    @Transactional
    public User create(User user){
        return userRepository.save(user);
    }

    @Override
    public User findByEmail(String email) {
        return userRepository.findByEmail(email)
                .orElseThrow(() -> new CustomException("Пользователь с таким email не существует"));
    }

    @Override
    public User getUserById(Long id) {
        return userRepository.findById(id).orElseThrow(() -> new RuntimeException("Пользователя с таким id нет"));
    }



}
